# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/colors/": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Create a Color",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Colors"
                ],
                "summary": "Create a Color",
                "parameters": [
                    {
                        "description": "Create a Color",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.CreateColorRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Color response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.ColorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/colors/get-by-filter": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get Colors",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Colors"
                ],
                "summary": "Get Colors",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.PaginationInputWithFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Color response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.PagedList-github_com_amirhasanpour_bama-clone-web-api_src_api_dto_ColorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/colors/{id}": {
            "get": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get a Color",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Colors"
                ],
                "summary": "Get a Color",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Color response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.ColorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Update a Color",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Colors"
                ],
                "summary": "Update a Color",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update a Color",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.UpdateColorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Color response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.ColorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Delete a Color",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Colors"
                ],
                "summary": "Delete a Color",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_amirhasanpour_bama-clone-web-api_src_api_dto.ColorResponse": {
            "type": "object",
            "properties": {
                "hexCode": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_amirhasanpour_bama-clone-web-api_src_api_dto.CreateColorRequest": {
            "type": "object",
            "properties": {
                "hexCode": {
                    "type": "string",
                    "maxLength": 7,
                    "minLength": 7
                },
                "name": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 3
                }
            }
        },
...
        "github_com_amirhasanpour_bama-clone-web-api_src_api_dto.Filter": {
            "type": "object",
            "properties": {
                "filterType": {
                    "description": "text number",
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "type": {
                    "description": "contains notContains equals notEqual startsWith lessThan lessThanOrEqual greaterThan greaterThanOrEqual inRange endsWith",
                    "type": "string"
                }
            }
        },
...
        "github_com_amirhasanpour_bama-clone-web-api_src_api_dto.PagedList-github_com_amirhasanpour_bama-clone-web-api_src_api_dto_ColorResponse": {
            "type": "object",
            "properties": {
                "hasNextPage": {
                    "type": "boolean"
                },
                "hasPreviousPage": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.ColorResponse"
                    }
                },
                "pageNumber": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                },
                "totalRows": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_amirhasanpour_bama-clone-web-api_src_api_dto.PaginationInputWithFilter": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.Filter"
                    }
                },
                "pageNumber": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "sort": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.Sort"
                    }
                }
            }
        },
...
        "github_com_amirhasanpour_bama-clone-web-api_src_api_dto.Sort": {
            "type": "object",
            "properties": {
                "colId": {
                    "type": "string"
                },
                "sort": {
                    "type": "string"
                }
            }
        },
...
        "github_com_amirhasanpour_bama-clone-web-api_src_api_dto.UpdateColorRequest": {
            "type": "object",
            "properties": {
                "hexCode": {
                    "type": "string",
                    "maxLength": 7,
                    "minLength": 7
                },
                "name": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 3
                }
            }
        },
...
        "github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "result": {},
                "resultCode": {
                    "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.ResultCode"
                },
                "success": {
                    "type": "boolean"
                },
                "validationErrors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_validations.ValidationError"
                    }
                }
            }
        },
        "github_com_amirhasanpour_bama-clone-web-api_src_api_helper.ResultCode": {
            "type": "integer",
            "enum": [
                0,
                40001,
                40101,
                40301,
                40401,
                42901,
                42902,
                50001,
                50002
            ],
            "x-enum-varnames": [
                "Success",
                "ValidationError",
                "AuthError",
                "ForbiddenError",
                "NotFoundError",
                "LimiterError",
                "OtpLimiterError",
                "CustomRecovery",
                "InternalError"
            ]
        },
        "github_com_amirhasanpour_bama-clone-web-api_src_api_validations.ValidationError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "property": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AuthBearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}